requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
joblib==1.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==3.0.1